requirements.txt
asgiref
certifi
charset-normalizer
coverage
diff-match-patch
Django
django-import-export
django-leaflet
django-widget-tweaks
djangorestframework
idna
pillow
psycopg2-binary
requests
sqlparse
tablib
typing_extensions
urllib3
